separate update view and move consensus to a split package
